:sparkles: 이벤트 생성하기 버튼 클릭 시 group id 넘겨주도록 세팅

+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/ui/main/grouphome/navigation/GroupHomeNavigation.kt
@@ -13,7 +13,7 @@ fun NavController.navigateGroupHome() {
 
 fun NavGraphBuilder.groupHomeNavGraph(
     onClickGroupDetail: (id: Long) -> Unit,
-    onClickEventMake: () -> Unit,
+    onClickEventMake: (Long) -> Unit,
     onClickMyPage: () -> Unit,
     onClickGroupMake: () -> Unit,
     onClickGroupEnter: () -> Unit,
